Comment out dead aosp lowbattery sound code
Notification.Bulder.setSound is now deprecated and NotificationChannels.ALERTS
sound will always override it, so attachLowBatterySound() is just dead
code.
To replicate: just wait for low battery (5%) and you'll get the default
Alerts channel sound (that anyway is customizable into SystemUI
notifications settings) instead of the lowbattery one.

An alternative is to create a new notification channel for the
lowbattery notification, but i'm more lazy than lazy Google :V
Let's see what they'll eventually do with this in the future.

Change-Id: I78c421ef5ef16179f582f350d0a07349e7a8d286